[Refactor] Search폴더를 home폴더 아래로 이동
