+++ src/App.tsx
@@ -2,8 +2,8 @@ import './App.css';
 import CustomMain from './pages/custom/custom-main/custom';
 import UserNickname from './pages/onboarding/UserNickname';
 import UserPassword from './pages/onboarding/UserPassword';
-import Home from './pages/main/home/Home';       
-import SearchDrink from './pages/search/searchDrink/SearchDrink'; 
+import Home from './pages/main/home/Home';
+import SearchDrink from './pages/main/search/searchDrink/SearchDrink';
 import Mypage from './pages/mypage/Mypage';
 import LoginButton from './pages/onboarding/Kakaotalk/LoginButton';
 // import SearchDrink from './pages/search/searchDrink/SearchDrink';
