+++ src/pages/main/home/home-today/DailyIntake.tsx
@@ -2,12 +2,19 @@ import React from 'react';
 
 const DailyIntake = () => {
   return (
-    <div className="w-full flex flex-col p-5 border shadow-lg rounded-[20px]"> {/* ring-1 ring-[#F4F4F4] */}
+    <div
+      className="w-full flex flex-col p-5 border shadow-lg rounded-[20px]"
+      style={{
+        boxShadow: '0 0px 15px rgba(0, 0, 0, 0.1)', // X축 0, Y축 10px
+      }}
+    >
+      {' '}
+      {/* ring-1 ring-[#F4F4F4] */}
       <div className="flex justify-between">
         <div className="flex flex-col">
           <div className="flex items-baseline">
-            <div className='text-xs'>적정 섭취량&nbsp;&nbsp;</div>
-            <div className='text-sm'>25g</div>
+            <div className="text-xs">적정 섭취량&nbsp;&nbsp;</div>
+            <div className="text-sm">25g</div>
           </div>
           <div className="mt-3 text-xl">
             아직 당 00g을<br></br>더 마실 수 있어요!
@@ -22,7 +29,7 @@ const DailyIntake = () => {
           <div className="text-center">당 섭취량</div>
           <div className="text-center">00g</div>
         </div>
-        <div className='w-[1.5px] h-10 bg-[#F4F4F4]'></div>
+        <div className="w-[1.5px] h-10 bg-[#F4F4F4]"></div>
         <div className="flex-1">
           <div className="text-center">칼로리</div>
           <div className="text-center">000kcal</div>
