+++ src/pages/main/home/home-week/WeeklyInfo.tsx
@@ -14,8 +14,8 @@ const WeeklyInfo = () => {
       <div className="w-full h-[15px] mt-7 bg-[#F4F4F4]"></div>
 
       {/* 주간 음료수 당 통계 */}
-      <div className="flex justify-between w-[calc(100%-48px)] mt-6 mb-5">
-        <div className="text-[18px]">주간 음료수 당 통계</div>
+      <div className="flex justify-between w-[calc(100%-48px)] my-6">
+        <div className="text-[17px]">주간 음료수 당 통계</div>
       </div>
 
       {/* 섭취량 통계표 */}
