+++ src/pages/main/home/home-week/WeeklyIntake.tsx
@@ -1,11 +1,17 @@
-import React from 'react'
+import React from 'react';
 
 const WeeklyIntake = () => {
   return (
-    <div className="w-full h-[225px] flex flex-col p-5 border shadow-lg rounded-3xl"> {/* ring-4 ring-[#F4F4F4] */}
-      
+    <div
+      className="w-full h-[225px] flex flex-col p-5 border shadow-lg rounded-3xl"
+      style={{
+        boxShadow: '0 0px 15px rgba(0, 0, 0, 0.1)', // X축 0, Y축 10px
+      }}
+    >
+      {' '}
+      {/* ring-4 ring-[#F4F4F4] */}
     </div>
-  )
-}
+  );
+};
 
-export default WeeklyIntake
+export default WeeklyIntake;
